refactor(hickey): centralize iframe-preview URL contract in server module
After the previous commit moved the extension predicate into
`iframePreviewRoute.ts`, the URL builder was the last iframe-preview
piece left in `kolu-git/schemas.ts`. The client never calls
`buildIframePreviewUrl` — it receives a pre-built URL via
`FsReadFileOutput.url` and renders it. The contract is between two
server files (`surface.ts` emits, `index.ts` matches), both in the same
package as `iframePreviewRoute.ts` — not a wire contract requiring
cross-package placement.

Also extract `TERMINAL_FILE_ROUTE_BASE` (and `TERMINAL_FILE_ROUTE_FILE_SEGMENT`)
so the builder, the Hono `app.get` registration, and the request-time
`prefix` slice all derive from one constant. URL renames now touch one
place instead of three.
