+++ packages/server/src/index.ts
@@ -14,7 +14,12 @@ import pkg from "../package.json" with { type: "json" };
 import { getCacheControlHeader } from "./cacheControl.ts";
 import { startDiagnostics } from "./diagnostics.ts";
 import { serverHostname } from "./hostname.ts";
-import { resolvePreviewPath, serveResolvedFile } from "./iframePreviewRoute.ts";
+import {
+  resolvePreviewPath,
+  serveResolvedFile,
+  TERMINAL_FILE_ROUTE_BASE,
+  TERMINAL_FILE_ROUTE_FILE_SEGMENT,
+} from "./iframePreviewRoute.ts";
 import { ensureKoluRoot, shutdownCleanup } from "./koluRoot.ts";
 import { log } from "./log.ts";
 import { pwaIdentityForHostname } from "./pwaIdentity.ts";
@@ -152,33 +157,36 @@ app.get("/api/health", (c) => c.text("kolu"));
 
 // --- Iframe preview file route ---
 // Serves repo files referenced by `FsReadFileOutput.kind === "binary"`.
-// URL contract lives in `kolu-git/schemas`'s `buildIframePreviewUrl`.
+// URL contract (base + builder + parser) all lives in `iframePreviewRoute.ts`.
 // Registered before the static-serve catch-all so production builds don't
 // shadow this route with `serveStatic`'s `/*` matcher.
-app.get("/api/terminals/:terminalId/file/*", async (c) => {
-  const terminalId = c.req.param("terminalId");
-  const prefix = `/api/terminals/${terminalId}/file/`;
-  // Use the raw decoded path from req.url to keep our own URL-decoding
-  // pipeline (see `resolvePreviewPath`) in charge of segment parsing —
-  // anything Hono pre-decodes here would bypass the guard.
-  const rawTail = c.req.path.startsWith(prefix)
-    ? c.req.path.slice(prefix.length)
-    : "";
+app.get(
+  `${TERMINAL_FILE_ROUTE_BASE}/:terminalId/${TERMINAL_FILE_ROUTE_FILE_SEGMENT}/*`,
+  async (c) => {
+    const terminalId = c.req.param("terminalId");
+    const prefix = `${TERMINAL_FILE_ROUTE_BASE}/${terminalId}/${TERMINAL_FILE_ROUTE_FILE_SEGMENT}/`;
+    // Use the raw decoded path from req.url to keep our own URL-decoding
+    // pipeline (see `resolvePreviewPath`) in charge of segment parsing —
+    // anything Hono pre-decodes here would bypass the guard.
+    const rawTail = c.req.path.startsWith(prefix)
+      ? c.req.path.slice(prefix.length)
+      : "";
 
-  const term = getTerminal(terminalId);
-  const repoRoot = term?.meta.git?.repoRoot;
-  if (!repoRoot) return c.text("terminal has no repo", 404);
+    const term = getTerminal(terminalId);
+    const repoRoot = term?.meta.git?.repoRoot;
+    if (!repoRoot) return c.text("terminal has no repo", 404);
 
-  const res = await serveResolvedFile(resolvePreviewPath(repoRoot, rawTail));
-  // `Buffer` (subclass of `Uint8Array<ArrayBufferLike>`) is a runtime-valid
-  // `BodyInit` but the DOM-typed lib.dom.d.ts narrows `BodyInit` to
-  // `Uint8Array<ArrayBuffer>` — the unions don't align in TS even though
-  // node-server forwards the buffer unchanged. Cast at the boundary.
-  return new Response(res.body as BodyInit, {
-    status: res.status,
-    headers: res.headers,
-  });
-});
+    const res = await serveResolvedFile(resolvePreviewPath(repoRoot, rawTail));
+    // `Buffer` (subclass of `Uint8Array<ArrayBufferLike>`) is a runtime-valid
+    // `BodyInit` but the DOM-typed lib.dom.d.ts narrows `BodyInit` to
+    // `Uint8Array<ArrayBuffer>` — the unions don't align in TS even though
+    // node-server forwards the buffer unchanged. Cast at the boundary.
+    return new Response(res.body as BodyInit, {
+      status: res.status,
+      headers: res.headers,
+    });
+  },
+);
 
 // --- Dynamic PWA manifest (includes hostname) ---
 app.get("/manifest.webmanifest", (c) => {
