+++ packages/server/src/surface.ts
@@ -40,7 +40,6 @@ import { contract } from "kolu-common/contract";
 import { TerminalNotFoundError } from "kolu-common/errors";
 import { surface } from "kolu-common/surface";
 import {
-  buildIframePreviewUrl,
   fsListAllOutputEqual,
   type FsReadFileOutput,
   fsReadFileOutputEqual,
@@ -55,7 +54,10 @@ import {
   subscribeFileChange,
   subscribeRepoChange,
 } from "kolu-git";
-import { isIframePreviewable } from "./iframePreviewRoute.ts";
+import {
+  buildIframePreviewUrl,
+  isIframePreviewable,
+} from "./iframePreviewRoute.ts";
 import { log } from "./log.ts";
 import { publisher } from "./publisher.ts";
 import { cancelPendingAutosave, getSavedSession } from "./session.ts";
